feat(zudo-doc): S7 — doc-page rendering internals → package factories (#2352)
Move 7 doc-page rendering modules (`_doc-pager`, `_doc-body-end`,
`_doc-metainfo-area`, `_doc-tags-area`, `_doc-history-area`,
`_doc-content-header`, `_doc-page-renderer`) from `pages/lib/` into
`@takazudo/zudo-doc` as injected-context factories. Also parameterize
`src/utils/github.ts` → `./github-helpers` and `src/utils/tags.ts` →
`./tag-helpers` as pure functions. Host stubs become thin wrappers;
create-zudo-doc base template updated in lockstep. `resolvePageTags`
builds the vocabulary index once per call (O(N) instead of O(N×M)).
