+++ src/utils/github.ts
@@ -1,19 +1,22 @@
-import { settings } from "@/config/settings";
+// Host thin-stub — see @takazudo/zudo-doc/github-helpers (epic #2344, S7).
+//
+// `buildGitHubRepoUrl` and `buildGitHubSourceUrl` are now parameterized pure
+// functions in the package; this module wraps them with the host's
+// `settings.githubUrl` singleton.
 
-function trimTrailingSlash(url: string): string {
-  return url.replace(/\/+$/, "");
-}
+import { settings } from "@/config/settings";
+import {
+  buildGitHubRepoUrl as _buildGitHubRepoUrl,
+  buildGitHubSourceUrl as _buildGitHubSourceUrl,
+} from "@takazudo/zudo-doc/github-helpers";
 
 export function buildGitHubRepoUrl(): string | null {
-  if (!settings.githubUrl) return null;
-  return trimTrailingSlash(settings.githubUrl as string);
+  return _buildGitHubRepoUrl(settings.githubUrl);
 }
 
 export function buildGitHubSourceUrl(
   contentDir: string,
   entryId: string,
 ): string | null {
-  const repoUrl = buildGitHubRepoUrl();
-  if (!repoUrl) return null;
-  return `${repoUrl}/blob/HEAD/${contentDir}/${entryId}`;
+  return _buildGitHubSourceUrl(settings.githubUrl, contentDir, entryId);
 }
